docs: add security section
This adds a 'Security' section, the intent of this page is to provide details about security features and design choices.

For now, it only document the IP address filtering feature and other hardening we have around outgoing traffic. In the future, we may want to expand it further to provide details such as rate-limiting, and Cloud deployment best-practices.