docs: add security section (#1405)
This adds a 'Security' section, the intent of this page is to provide details about security features and design choices.

For now, it only documents the IP address filtering feature and other hardening we have around outgoing traffic. In the future, we may want to expand it further to provide details such as rate-limiting, and self-hosting best-practices.